Update invoice documentations for REFUND_ADJ.
Invoice adjustments are a legacy feature, and invoice items with type
REFUND_ADJ was removed from Kill Bill in 0.17.3 or later. I'm updating
the documentation to remove the references to REFUND_ADJ, and remove
the section that mentions invoice adjustments (since only invoice
item adjustments are supported now).

I also updated the section about calculating the amount on the invoice,
since I notice the "amount" on the invoice does change in the prescence
of REPAIR_ADJ and ITEM_ADJ. Please let me know if that's right.